Addressing issue #65

+++ src/vlog/forward/ruleexecplan.cpp
@@ -16,7 +16,7 @@ void RuleExecutionPlan::checkIfFilteringHashMapIsPossible(const Literal &head) {
     }
 
     bool differentVar = false;
-    for (uint8_t i = 0; i < head.getTupleSize(); ++i) {
+    for (int i = 0; i < head.getTupleSize(); ++i) {
         VTerm th = head.getTermAtPos(i);
         VTerm tl = lastLit->getTermAtPos(i);
         if (!th.isVariable()) {
@@ -69,7 +69,7 @@ void RuleExecutionPlan::calculateJoinsCoordinates(const std::vector<Literal> &he
     std::map<Var_t,std::vector<uint8_t>> variablesNeededForHead;
     uint32_t countVars = 0;
     for (auto &headLiteral : heads) {
-        for (uint8_t headPos = 0; headPos < headLiteral.getTupleSize(); ++headPos) {
+        for (int headPos = 0; headPos < headLiteral.getTupleSize(); ++headPos) {
             const VTerm headTerm = headLiteral.getTermAtPos(headPos);
             if (headTerm.isVariable()) {
                 variablesNeededForHead[headTerm.getId()].push_back(countVars + headPos);
@@ -78,7 +78,7 @@ void RuleExecutionPlan::calculateJoinsCoordinates(const std::vector<Literal> &he
         countVars += headLiteral.getTupleSize();
     }
 
-    for (uint8_t i = 0; i < plan.size(); ++i) {
+    for (int i = 0; i < plan.size(); ++i) {
         const Literal *currentLiteral = plan[i];
 
         std::vector<std::pair<uint8_t, uint8_t>> jc;
@@ -100,7 +100,7 @@ void RuleExecutionPlan::calculateJoinsCoordinates(const std::vector<Literal> &he
             //No need to store any new variable. Just copy the old ones in the head
             //if the variables in "existingVariables" are needed for the head, then
             //save in "pf" [_idx_in_term_list_in_head_,_idx_in_existingVariables_]
-            for (uint8_t m = 0; m < existingVariables.size(); ++m) {
+            for (int m = 0; m < existingVariables.size(); ++m) {
                 if (variablesNeededForHead.count(existingVariables[m])) {
                     auto p = variablesNeededForHead.find(existingVariables[m]);
                     for (auto &el : p->second) {
@@ -110,11 +110,11 @@ void RuleExecutionPlan::calculateJoinsCoordinates(const std::vector<Literal> &he
             }
         } else {
             //copy only the ones that will be used later on
-            for (uint8_t j = 0; j < existingVariables.size(); ++j) {
+            for (int j = 0; j < existingVariables.size(); ++j) {
                 // first check if the variable existingVariables[j] is needed in the future
                 bool isVarNeeded = false;
                 //Check in the rest of the body if the variable is mentioned
-                for (uint8_t m = i + 1; m < plan.size(); ++m) {
+                for (int m = i + 1; m < plan.size(); ++m) {
                     if (plan[m]->containsVariable(existingVariables[j])){
                         isVarNeeded = true;
                         break;
@@ -138,7 +138,7 @@ void RuleExecutionPlan::calculateJoinsCoordinates(const std::vector<Literal> &he
         //Put in join coordinates between the previous and the current literal
         uint8_t litVars = 0;
         std::set<Var_t> processed; // This set is used to avoid repeated variables in the literal.
-        for (uint8_t x = 0; x < currentLiteral->getTupleSize(); ++x) {
+        for (int x = 0; x < currentLiteral->getTupleSize(); ++x) {
             const VTerm t = currentLiteral->getTermAtPos(x);
 
             if (t.isVariable()) {
@@ -162,7 +162,7 @@ void RuleExecutionPlan::calculateJoinsCoordinates(const std::vector<Literal> &he
                     // first occurrence.
                     bool isVariableNeeded = false;
                     //Check next literals
-                    for (uint8_t m = i + 1; m < plan.size(); ++m) {
+                    for (int m = i + 1; m < plan.size(); ++m) {
                         if (plan[m]->containsVariable(t.getId())){
                             isVariableNeeded = true;
                             break;
